requirements.txt
Flask
requests
PyPDF2
Pillow
pytesseract
markdownify
beautifulsoup4
python-dotenv
markdown
